requirements.txt
farm-haystack==1.25.5
numpy>=1.26.0
scipy>=1.10.0
scikit-learn>=1.3.0
pandas>=2.0.0
joblib
click
rich
langgraph==0.4.0
requests>=2.31.0
pytest>=8.0.0
